requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.1
colorama==0.4.4
distlib==0.3.1
Django==3.2.5
djangorestframework==3.12.4
filelock==3.0.12
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
pylint==2.7.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.2
wrapt==1.12.1
